Added a message to get the node list
However, the O2 library is not working properly with URLs with %20 or ?

+++ src/com/egg-software/ubuntuone-qt-files/dto/NodeInfoDTO.h
@@ -49,7 +49,7 @@ public:
     // Only for type FILE
     bool isPublic; /// If the file is public or not
     QString publicUrl; /// If the file is public, the public URL
-    qint64 size; /// File size in bytes
+    int size; /// File size in bytes
     QString lastModified; /// DD/MM/YYYY
 };
